Fix style application on 01-cascade-fix page
Add, remove, and reorder selectors and declaration blocks as needed to
achieve desired webpage styles for each element.

This fix did not include any changes to the pre-existing style
declarations themselves, only to the order of those declarations, as
well as changes to selectors. In some cases multiple edits were made
to a selector/block combination for cascade redundancy.