Change header display type to flexbox
Set display type for header to flex with items aligned center and
content justified with space-between so that all children are vertically
aligned center, left links stay on left edge, and right links stay on
right edge. Add padding of 8px.
Set display type for ul elements to flex so they follow same rules and
manually set margin and padding for ul to 0 to remove their defaults.
Set gap to 10px so each li and child anchor are clearly delineated.
This layout is maintained even as page is shrunk or expanded.