Change header to be flexible in middle only
Nest logo and list elements in new div with class of 'left-content'.
Nest button and proifle-pic elements in new div with class of 'right-
content'.

Add padding to header element and change display type to flexbox with
children justified with space-between and items aligned center.
Change display type of new dev elements to flexbox with a gap of 16px
and children aligned center.
Change display type of ul to flexbox with gap of 16px and default
margin/padding overwritten to 0's.

The outcome of this should be a header with left-most elements and
right-most elements locked against the edges of the page as it grows
and shrinks along horizontal axis.